Moved try catch block for likes handling and blog removing to the App.js side so that the tests work properly, also fixed the blog creator to show without page refresh
